♻️ user 폴더 아래에 사용 가능한 페이지들을 옮겨 404 에러 발생시 gnb 및 footer 생기지 않도록 변경하고 수정된 경로에 따른 각 파일들의 router.push나 redirect, Link의 href 경로 조정
