--- src/app/filterTest/page.tsx
@@ -1,8 +1,0 @@
-import AdvancedFilterComponent from "@/components/common/AdvancedFilter/AdvancedFilterComponent";
-import React from "react";
-
-const FilterTestPage = () => {
-  return <>{/* <AdvancedFilterComponent /> */}</>;
-};
-
-export default FilterTestPage;
