+++ src/app/layout.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import Footer from "@/components/common/Footer";
-import GNB from "@/components/common/GNB";
 
 export const metadata: Metadata = {
   title: "더 줄게",
@@ -15,9 +13,7 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <body className="tab:px-8 mob:px-3">
-        <GNB />
-        <main>{children}</main>
-        <Footer />
+        {children}
         <div id="modal-root" />
       </body>
     </html>
