+++ src/components/employer/EmployerNoticeForm.tsx
@@ -142,7 +142,7 @@ const EmployerNoticeForm = ({}: EmployerNoticeForm) => {
 
   const handleModalClose = () => {
     setShowModal(false);
-    router.push(`/employer/notice/${objectNoticeId}`);
+    router.push(`/user/employer/notice/${objectNoticeId}`);
   };
 
   return (
